+++ src/main/resources/data.sql
@@ -3,8 +3,8 @@ DROP TABLE IF EXISTS vehicle;
 CREATE TABLE vehicle (
   id VARCHAR(50) PRIMARY KEY,
   name VARCHAR(100) NOT NULL,
-  x INT NOT NULL,
-  y INT NOT NULL,
+  x DOUBLE NOT NULL,
+  y DOUBLE NOT NULL,
   licencePlate VARCHAR(50) NOT NULL,
   range INT NOT NULL,
   batteryLevel INT NOT NULL,
